Use fast, static lookup table for most common key bindings
The old PointerArray based lookup method is now only used for uncommon
bindings, multi-key chains, Unicode keys, etc.

The new lookup table has only 432 indices but covers all of the default
bindings, except for one:

    bind 'M-^?' 'erase-word -s'

+++ src/bind.h
@@ -1,11 +1,12 @@
 #ifndef BIND_H
 #define BIND_H
 
+#include <stddef.h>
 #include "key.h"
 
 void add_binding(const char *keys, const char *command);
 void remove_binding(const char *keys);
 void handle_binding(KeyCode key);
-int nr_pressed_keys(void);
+size_t nr_pressed_keys(void);
 
 #endif
